+++ plugin.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-"""QGIS plugin bootstrap for VertoBridge Italia."""
+"""QGIS plugin bootstrap for GeoBridgeIT."""
 
 from __future__ import annotations
 
@@ -7,25 +7,25 @@ import os
 
 from qgis.PyQt.QtGui import QIcon
 
-from .dialog import VertoDialog
+from .dialog import GeoBridgeDialog
 from .qt_compat import QAction
 
 
-class VertoBridgeItaliaPlugin:
+class GeoBridgeITPlugin:
     def __init__(self, iface):
         self.iface = iface
         self.action = None
         self.dialog = None
-        self.menu_name = "&VertoBridge Italia"
+        self.menu_name = "&GeoBridgeIT"
 
     def initGui(self):
         icon_path = os.path.join(os.path.dirname(__file__), "icon.svg")
         self.action = QAction(
-            QIcon(icon_path), "VertoBridge Italia", self.iface.mainWindow()
+            QIcon(icon_path), "GeoBridgeIT", self.iface.mainWindow()
         )
-        self.action.setObjectName("vertobridge_italia_action")
+        self.action.setObjectName("geobridgeit_action")
         self.action.setToolTip(
-            "Client QGIS non ufficiale per conversioni tramite API Verto Online"
+            "Client QGIS non ufficiale per conversioni tramite API servizio API IGM"
         )
         self.action.triggered.connect(self.run)
         self.iface.addToolBarIcon(self.action)
@@ -39,7 +39,7 @@ class VertoBridgeItaliaPlugin:
         self.dialog = None
 
     def run(self):
-        self.dialog = VertoDialog(self.iface)
+        self.dialog = GeoBridgeDialog(self.iface)
         if hasattr(self.dialog, "exec"):
             self.dialog.exec()
         else: