+++ CHANGELOG.md
@@ -6,8 +6,8 @@ Il formato è basato su [Keep a Changelog](https://keepachangelog.com/it-IT/1.0.
 
 ## [1.0.0] - 2026-06-08
 ### Aggiunto
-- Prima release ufficiale (Stabile) di VertoBridge Italia.
-- Supporto per la conversione planimetrica di coordinate singole e layer vettoriali tramite API Verto Online IGM.
+- Prima release ufficiale (Stabile) di GeoBridgeIT.
+- Supporto per la conversione planimetrica di coordinate singole e layer vettoriali tramite API servizio API IGM IGM.
 - Compatibilità completa con QGIS 3 (Qt5) e QGIS 4 (Qt6).
 - Integrazione UI "Cyber Mint".
 - Avvisi e disclaimer legali in conformità all'uso delle API pubbliche IGM.